refactor(http_funcs): carve up test files ahead of adding coverage
